// src/types/Topic.ts
//
export enum Topics {
  RUN_COMMAND = 'RUN_COMMAND',
  QUERY = 'QUERY',
  LOGIN = 'LOGIN',
  LOGOUT = 'LOGOUT',
  LOGIN_FAILED = 'LOGIN_FAILED',
  SUCCESS = 'SUCCESS',
  
  GET_WALLETS = 'GET_WALLETS',
  GET_SUBNETS = 'GET_SUBNETS',
  GET_DEFAULT_WALLET = 'GET_DEFAULT_WALLET',
  CREATE_SUBNET = 'CREATE_SUBNET',
  ADD_WALLET = 'ADD_WALLET',
  ADD_PRIVATE_KEY = 'ADD_PRIVATE_KEY',
  SET_DEFAULT_WALLET = 'SET_DEFAULT_WALLET'
}